AMDGPU/SI: Use Constant instead of TargetConstant for some MUBUF complex patterns
When an operand in a pattern is defined with imm:$offset then it will only
match Constant nodes.  When an operand is define with ValueType:$offset, then
it will match Constant and TargetConstant nodes, but only the TargetConstant
nodes will be folded into the instruction.

We should use Constant nodes for immediate operands that are always immediates
and TargeConstant nodes for immediate operands that could be either an
immediate or a register.